Split QueryPerformance.cs base into PlanXml/Heatmap/Comparison partials
Behavior-neutral mechanical partial-class split. Pulls three concern
clusters out of the tail of DatabaseService.QueryPerformance.cs into new
sibling partials of the same `public partial class DatabaseService`:

- DatabaseService.QueryPerformance.PlanXml.cs: the 5 on-demand plan-XML
  fetch methods (GetQueryStorePlanXmlByCollectionIdAsync,
  GetProcedureStatsPlanXmlByCollectionIdAsync,
  GetQueryStatsPlanXmlByCollectionIdAsync, GetPlanXmlByQueryHashAsync,
  GetProcedurePlanXmlBySqlHandleAsync).
- DatabaseService.QueryPerformance.Heatmap.cs: GetQueryHeatmapAsync plus
  its heatmap-only helpers (GetHeatmapMetricExpr, HeatmapDurationLabels,
  HeatmapCountLabels).
- DatabaseService.QueryPerformance.Comparison.cs: the 3 comparison methods
  (GetQueryStatsComparisonAsync, GetProcedureStatsComparisonAsync,
  GetQueryStoreComparisonAsync).

Members moved verbatim (no reformatting). Base file shrinks 1200 -> 580
lines; no call-site changes. Base diff is a pure deletion (0 additions,
620 deletions), confirming kept methods are untouched.

Build: `dotnet build PerformanceMonitor.sln -c Debug` -> Build succeeded,
0 warnings, 0 errors.
Tests: `dotnet test Dashboard.Tests` -> Passed: 488, Failed: 0.
